test: match account level smallint type

+++ supabase/tests/database/accounts_integrity.test.sql
@@ -17,7 +17,7 @@ values (
 
 select is(
   (select level from public.accounts where id = '61000000-0000-0000-0000-000000000001'),
-  1,
+  1::smallint,
   'root account level is derived by the database'
 );
 
@@ -39,7 +39,7 @@ select lives_ok(
 
 select is(
   (select level from public.accounts where id = '61000000-0000-0000-0000-000000000002'),
-  2,
+  2::smallint,
   'child level is derived from its parent'
 );
